// src/app/models/user.ts
export class User {
    uid?: string;
    nickname?: string;
    username?: string;
    email?: string;
    password?: string;
    avatar?: string;
    institute?: string;
    role?: string;
    gender?: string;
    status?: string;
    level?: number;
    exp?: number;
    money?: number;
    joindate?: Date;
    phonenumber?: string;
    lastseen?: Date;
    usechar?: string;
    usewall?: string;
    usedecor?: string;
    title?: string;
    idchar?: string;
}